Funciona todo a falta de hacer buscador

+++ src/app/features/admin-productos/admin-productos.component.ts
@@ -17,6 +17,7 @@ import { CommonModule } from '@angular/common';
 })
 export class AdminProductosComponent implements OnInit {
   productos: ProductoDTO[] = [];
+  modalCrearVisible: boolean = false;
 
   entregas = [
     { label: 'RECOGIDA', value: 0 },
@@ -52,14 +53,23 @@ export class AdminProductosComponent implements OnInit {
 
   cargarProductos(): void {
     this.productoService.getProductos().subscribe(data => {
-      this.productos = data;
+      this.productos = data; // Asignar directamente los datos del backend
     });
+  }
+
+  mostrarModalCrear(): void {
+    this.resetForm();
+    this.modalCrearVisible = true;
+  }
+
+  cerrarModalCrear(): void {
+    this.modalCrearVisible = false;
   }
 
   agregarProducto(): void {
     this.productoService.crearProducto(this.nuevoProducto).subscribe(() => {
       this.cargarProductos();
-      this.resetForm();
+      this.cerrarModalCrear();
     });
   }
 
@@ -81,9 +91,11 @@ export class AdminProductosComponent implements OnInit {
   }
 
   eliminarProducto(id: number): void {
-    this.productoService.eliminarProducto(id).subscribe(() => {
-      this.cargarProductos();
-    });
+    if (confirm('¿Estás seguro de que deseas eliminar este producto?')) {
+      this.productoService.eliminarProducto(id).subscribe(() => {
+        this.cargarProductos();
+      });
+    }
   }
 
   resetForm(): void {
